+++ packages/server/database/types/Meeting.ts
@@ -44,6 +44,7 @@ export default abstract class Meeting {
   summary?: string
   sentimentScore?: number
   usedReactjis?: Record<string, number>
+  slackTs?: string
 
   constructor(input: Input) {
     const {