+++ packages/server/graphql/public/mutations/updateUserProfile.ts
@@ -6,7 +6,6 @@ import updateUser from '../../../postgres/queries/updateUser'
 import {analytics} from '../../../utils/analytics/analytics'
 import {getUserId, isAuthenticated} from '../../../utils/authorization'
 import publish from '../../../utils/publish'
-import segmentIo from '../../../utils/segmentIo'
 import standardError from '../../../utils/standardError'
 import {MutationResolvers} from '../resolverTypes'
 
@@ -63,12 +62,10 @@ const updateUserProfile: MutationResolvers['updateUserProfile'] = async (
   const user = await dataLoader.get('users').loadNonNull(userId)
   if (normalizedPreferredName) {
     analytics.accountNameChanged(userId, normalizedPreferredName)
-    segmentIo.identify({
+    analytics.identify({
       userId,
-      traits: {
-        email: user.email,
-        name: normalizedPreferredName
-      }
+      email: user.email,
+      name: normalizedPreferredName
     })
   }