+++ packages/server/graphql/public/types/ReflectTemplate.ts
@@ -1,8 +1,9 @@
 import ms from 'ms'
-import errorFilter from '../../errorFilter'
-import {ReflectTemplateResolvers} from '../resolverTypes'
 import {getUserId} from '../../../utils/authorization'
+import errorFilter from '../../errorFilter'
 import {DataLoaderWorker} from '../../graphql'
+import isValid from '../../isValid'
+import {ReflectTemplateResolvers} from '../resolverTypes'
 
 const POPULAR_RETROS = [
   'workingStuckTemplate',
@@ -13,15 +14,18 @@ const POPULAR_RETROS = [
   'fourLsTemplate'
 ]
 
-const getLastUsedDateForTeams = async (
-  teams: string[],
+const getLastUsedAtForTeams = async (
+  teamIds: string[],
   templateId: string,
   dataLoader: DataLoaderWorker
 ) => {
-  return (await dataLoader.get('retroTemplateLastUsedByTeam').loadMany(teams))
-    .filter(errorFilter)
-    .map((templateLookup) => templateLookup[templateId] || new Date(0))
-    .reduce((dateA, dateB) => (dateA > dateB ? dateA : dateB), new Date(0))
+  const teamMeetingTemplates = await dataLoader.get('teamMeetingTemplateByTeamId').loadMany(teamIds)
+  const lastUsedAtsForTemplateId = teamMeetingTemplates
+    .filter(isValid)
+    .flat()
+    .filter((tmt) => tmt.templateId === templateId)
+    .map((row) => row.lastUsedAt.getTime())
+  return Math.max(...lastUsedAtsForTemplateId)
 }
 
 const ReflectTemplate: ReflectTemplateResolvers = {
@@ -49,12 +53,6 @@ const ReflectTemplate: ReflectTemplateResolvers = {
       subCategories.push('popular')
     }
 
-    // Recently Used
-    const dateLastUsedForTeam = await getLastUsedDateForTeams(authToken.tms, id, dataLoader)
-    if (dateLastUsedForTeam > new Date(Date.now() - ms('30d'))) {
-      subCategories.push('recentlyUsed')
-    }
-
     // Others are using in your org
     const viewerId = getUserId(authToken)
     const organizationUsers = await dataLoader.get('organizationUsersByUserId').load(viewerId)
@@ -64,17 +62,27 @@ const ReflectTemplate: ReflectTemplateResolvers = {
       .flat()
       .map((team) => team.id)
 
-    const dateLastUsedForOrg = await getLastUsedDateForTeams(
-      orgTeamIds.filter((orgId) => !authToken.tms.includes(orgId)),
-      id,
-      dataLoader
-    )
-    if (dateLastUsedForOrg > new Date(Date.now() - ms('30d'))) {
+    // The goal here is to issue all fetches to the DB at once, then we can parse out team membership
+    const {tms} = authToken
+    const [lastUsedAtOnTeam, lastUsedAtOnOrg] = await Promise.all([
+      getLastUsedAtForTeams(tms, id, dataLoader),
+      getLastUsedAtForTeams(
+        orgTeamIds.filter((teamId) => !tms.includes(teamId)),
+        id,
+        dataLoader
+      )
+    ])
+
+    if (lastUsedAtOnTeam > Date.now() - ms('30d')) {
+      subCategories.push('recentlyUsed')
+    }
+
+    if (lastUsedAtOnOrg > Date.now() - ms('30d')) {
       subCategories.push('recentlyUsedInOrg')
     }
 
     // Try these activities
-    if (dateLastUsedForTeam.getTime() === new Date(0).getTime()) {
+    if (lastUsedAtOnTeam <= 0) {
       subCategories.push('neverTried')
     }