+++ packages/server/graphql/queries/tasks.ts
@@ -17,7 +17,6 @@ import GraphQLISO8601Type from '../types/GraphQLISO8601Type'
 import {TaskConnection} from '../types/Task'
 import TaskStatusEnum, {TaskStatusEnumType} from '../types/TaskStatusEnum'
 import connectionFromTasks from './helpers/connectionFromTasks'
-import {getAccessibleTeamIdsForUser} from '../getAccessibleTeamIdsForUser'
 
 const getValidUserIds = async (
   userIds: null | string[],
@@ -124,7 +123,10 @@ export default {
 
     // if archived is true & no userId filter is provided, it should include tasks for ex-team members
     // under no condition should it show tasks for archived teams
-    const validTeamIds = await getAccessibleTeamIdsForUser(viewerId, teamIds, dataLoader)
+    const accessibleTeamIds = authToken.tms
+    const validTeamIds = teamIds
+      ? teamIds.filter((teamId: string) => accessibleTeamIds.includes(teamId))
+      : accessibleTeamIds
     const validUserIds = (await getValidUserIds(userIds, viewerId, validTeamIds, dataLoader)) ?? []
     // RESOLUTION
     const tasks = await dataLoader.get('userTasks').load({