+++ scripts/webpack/dev.servers.config.js
@@ -55,7 +55,6 @@ module.exports = {
   ],
   plugins: [
     new webpack.DefinePlugin({
-      __PROJECT_ROOT__: JSON.stringify(PROJECT_ROOT),
       __PRODUCTION__: false
     })
   ],